+++ BookServiceQA/Support classes/AssemblySetupFixture.cs
